# delivery/OPENAPI.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ResearchNavigator API",
    "description": "Evidence-grounded research direction and paper workspace.",
    "version": "2.2.2"
  },
  "paths": {
...
    "/api/admin/config-status": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Admin Config Status",
        "operationId": "admin_config_status_api_admin_config_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminConfigStatus"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/admin/runtime-config": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Get Admin Runtime Config",
        "operationId": "get_admin_runtime_config_api_admin_runtime_config_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminRuntimeConfig"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "system"
        ],
        "summary": "Update Admin Runtime Config",
        "operationId": "update_admin_runtime_config_api_admin_runtime_config_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminRuntimeConfig"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminRuntimeConfig"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/jobs": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "List Jobs",
        "operationId": "list_jobs_api_jobs_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/JobRead"
                  },
                  "type": "array",
                  "title": "Response List Jobs Api Jobs Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "system"
        ],
        "summary": "Create Job",
        "operationId": "create_job_api_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/jobs/{job_id}": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Get Job",
        "operationId": "get_job_api_jobs__job_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}/cancel": {
      "post": {
        "tags": [
          "system"
        ],
        "summary": "Cancel Job",
        "operationId": "cancel_job_api_jobs__job_id__cancel_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}/events": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Get Job Events",
        "operationId": "get_job_events_api_jobs__job_id__events_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobEventRead"
                  },
                  "title": "Response Get Job Events Api Jobs  Job Id  Events Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sources/status": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "List Source Status",
        "operationId": "list_source_status_api_sources_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SourceHealthRead"
                  },
                  "type": "array",
                  "title": "Response List Source Status Api Sources Status Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/sources/{source_name}/test": {
      "post": {
        "tags": [
          "system"
        ],
        "summary": "Test Source",
        "operationId": "test_source_api_sources__source_name__test_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "source_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceHealthRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AdminConfigStatus": {
        "properties": {
          "semantic_scholar_key_configured": {
            "type": "boolean",
            "title": "Semantic Scholar Key Configured"
          },
          "openalex_api_key_configured": {
            "type": "boolean",
            "title": "Openalex Api Key Configured"
          },
          "llm_key_configured": {
            "type": "boolean",
            "title": "Llm Key Configured"
          },
          "analysis_provider": {
            "type": "string",
            "title": "Analysis Provider"
          },
          "analysis_prompt_version": {
            "type": "string",
            "title": "Analysis Prompt Version"
          }
        },
        "type": "object",
        "required": [
          "semantic_scholar_key_configured",
          "openalex_api_key_configured",
          "llm_key_configured",
          "analysis_provider",
          "analysis_prompt_version"
        ],
        "title": "AdminConfigStatus"
      },
      "AdminRuntimeConfig": {
        "properties": {
          "source_health_timeout_seconds": {
            "type": "number",
            "maximum": 30.0,
            "minimum": 0.5,
            "title": "Source Health Timeout Seconds",
            "default": 5.0
          },
          "worker_max_attempts_default": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Worker Max Attempts Default",
            "default": 3
          }
        },
        "type": "object",
        "title": "AdminRuntimeConfig"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobCreate": {
        "properties": {
          "job_type": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Job Type"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          }
        },
        "type": "object",
        "required": [
          "job_type"
        ],
        "title": "JobCreate"
      },
      "JobEventRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "event_type": {
            "type": "string",
            "title": "Event Type"
          },
          "detail": {
            "additionalProperties": true,
            "type": "object",
            "title": "Detail"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "event_type",
          "detail",
          "created_at"
        ],
        "title": "JobEventRead"
      },
      "JobRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "job_type": {
            "type": "string",
            "title": "Job Type"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "result": {
            "additionalProperties": true,
            "type": "object",
            "title": "Result"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "attempt_count": {
            "type": "integer",
            "title": "Attempt Count"
          },
          "max_attempts": {
            "type": "integer",
            "title": "Max Attempts"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "cancelled_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancelled At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "terminal": {
            "type": "boolean",
            "title": "Terminal"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "job_type",
          "status",
          "payload",
          "result",
          "error",
          "attempt_count",
          "max_attempts",
          "started_at",
          "finished_at",
          "cancelled_at",
          "created_at",
          "terminal"
        ],
        "title": "JobRead"
      },
...
      "SourceHealthRead": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "configured": {
            "type": "boolean",
            "title": "Configured"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "checked_at": {
            "type": "string",
            "format": "date-time",
            "title": "Checked At"
          },
          "cooldown_until": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cooldown Until"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "name",
          "status",
          "enabled",
          "configured",
          "checked_at"
        ],
        "title": "SourceHealthRead"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}